refactor(process_tracker): move `PTRunner` to `scheduler::types` module
